user_service_app/user_service/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
pymysql
gunicorn
python-dotenv
cryptography>=41.0
requests
django-extensions>=3.2